backend/requirements.txt
# Backend dependencies
MetaTrader5>=5.0.45
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.26.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
